[MS-2]: Reverted changes in services.

+++ src/main/java/com/doit/wheels/dao/repositories/GenericRepository.java
@@ -3,4 +3,4 @@ package com.doit.wheels.dao.repositories;
 import com.doit.wheels.dao.entities.basic.AbstractModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenericRepository<T extends AbstractModel> extends JpaRepository<T, Long> {}
+public interface GenericRepository<T extends AbstractModel> extends JpaRepository<T, Integer> {}
